Backend/requirements.txt
fastapi
crewai
python-dotenv
motor
uvicorn
pydantic
